+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cubev",
-  "version": "0.2.1",
+  "version": "0.3.0",
   "scripts": {
     "dev": "vite",
     "build": "vite build"
