+++ src/App.vue
@@ -1,16 +1,19 @@
 <script setup>
 import Cubev from './Cubev.vue'
-let state1 = $ref({})
-let state2 = $ref({})
-
 import * as Code from '../plugins/Code.js'
 import * as Head from '../plugins/Head.js'
+import * as MarkdownRender from '../plugins/MarkdownRender.js'
+let state1 = $ref({})
+let state2 = $ref({
+  code: MarkdownRender.code,
+  markdown: '## Hello! \n\n This cubev uses plugin `MarkdownRender` \n\n \`\`\`js\nconsole.log(\'Welcome to Cubev!\')\`\`\`'
+})
 </script>
 
 <template>
   <cubev :state="state1" :plugins="[Code, Head]"></cubev>
   <hr style="margin: 50px;">
-  <cubev :state="state2"></cubev>
+  <cubev :state="state2" :plugins="[MarkdownRender]"></cubev>
 </template>
 
 <style>
